Further build changes

+++ rocket/rocket-model-cxxrtl.cpp
@@ -1,5 +1,5 @@
+#include "rocket-cxxrtl.cpp"
 #include "rocket-model.h"
-#include "rocket_cxxrtl.cpp"
 #include <backends/cxxrtl/cxxrtl_vcd.h>
 #include <iostream>
 
@@ -25,7 +25,7 @@ public:
 
   void vcd_dump(size_t cycle) override {
     // if (model_vcd)
-    // model_vcd->dump(static_cast<uunsigned64_t>(cycle));
+    // model_vcd->dump(static_cast<ulong unsigned64_t>(cycle));
   }
 
   void clock() override {
@@ -74,33 +74,33 @@ public:
     model.p_mem__axi4__0__w__ready.set<bool>(in.aw_ready);
     model.p_mem__axi4__0__w__ready.set<bool>(in.w_ready);
     model.p_mem__axi4__0__b__valid.set<bool>(in.b_valid);
-    model.p_mem__axi4__0__b__bits__id.set<unsigned>(in.b_id);
-    model.p_mem__axi4__0__b__bits__resp.set<unsigned>(in.b_resp);
+    model.p_mem__axi4__0__b__bits__id.set<uint32_t>(in.b_id);
+    model.p_mem__axi4__0__b__bits__resp.set<uint32_t>(in.b_resp);
     model.p_mem__axi4__0__r__ready.set<bool>(in.ar_ready);
     model.p_mem__axi4__0__r__valid.set<bool>(in.r_valid);
-    model.p_mem__axi4__0__r__bits__id.set<unsigned>(in.r_id);
-    model.p_mem__axi4__0__w__bits__data.set<unsigned>(in.r_data);
-    model.p_mem__axi4__0__w__bits__strb.set<unsigned>(in.r_resp);
-    model.p_mem__axi4__0__w__bits__last.set<unsigned>(in.r_last);
+    model.p_mem__axi4__0__r__bits__id.set<uint32_t>(in.r_id);
+    model.p_mem__axi4__0__w__bits__data.set<uint64_t>(in.r_data);
+    model.p_mem__axi4__0__w__bits__strb.set<uint32_t>(in.r_resp);
+    model.p_mem__axi4__0__w__bits__last.set<uint32_t>(in.r_last);
   }
 
   AxiOutputs get_mem() override {
     AxiOutputs out;
     out.aw_valid = model.p_mem__axi4__0__aw__valid.get<bool>();
-    out.aw_id = model.p_mem__axi4__0__aw__bits__id.get<unsigned>();
-    out.aw_addr = model.p_mem__axi4__0__aw__bits__addr.get<unsigned>();
-    out.aw_len = model.p_mem__axi4__0__aw__bits__len.get<unsigned>();
-    out.aw_size = model.p_mem__axi4__0__aw__bits__size.get<unsigned>();
+    out.aw_id = model.p_mem__axi4__0__aw__bits__id.get<uint32_t>();
+    out.aw_addr = model.p_mem__axi4__0__aw__bits__addr.get<uint32_t>();
+    out.aw_len = model.p_mem__axi4__0__aw__bits__len.get<uint32_t>();
+    out.aw_size = model.p_mem__axi4__0__aw__bits__size.get<uint32_t>();
     out.w_valid = model.p_mem__axi4__0__w__valid.get<bool>();
-    out.w_data = model.p_mem__axi4__0__w__bits__data.get<unsigned>();
-    out.w_strb = model.p_mem__axi4__0__w__bits__strb.get<unsigned>();
-    out.w_last = model.p_mem__axi4__0__w__bits__last.get<unsigned>();
+    out.w_data = model.p_mem__axi4__0__w__bits__data.get<uint64_t>();
+    out.w_strb = model.p_mem__axi4__0__w__bits__strb.get<uint32_t>();
+    out.w_last = model.p_mem__axi4__0__w__bits__last.get<uint32_t>();
     out.b_ready = model.p_mem__axi4__0__b__ready.get<bool>();
     out.ar_valid = model.p_mem__axi4__0__ar__valid.get<bool>();
-    out.ar_id = model.p_mem__axi4__0__ar__bits__id.get<unsigned>();
-    out.ar_addr = model.p_mem__axi4__0__ar__bits__addr.get<unsigned>();
-    out.ar_len = model.p_mem__axi4__0__ar__bits__len.get<unsigned>();
-    out.ar_size = model.p_mem__axi4__0__ar__bits__size.get<unsigned>();
+    out.ar_id = model.p_mem__axi4__0__ar__bits__id.get<uint32_t>();
+    out.ar_addr = model.p_mem__axi4__0__ar__bits__addr.get<uint32_t>();
+    out.ar_len = model.p_mem__axi4__0__ar__bits__len.get<uint32_t>();
+    out.ar_size = model.p_mem__axi4__0__ar__bits__size.get<uint32_t>();
     out.r_ready = model.p_mem__axi4__0__r__ready.get<bool>();
     return out;
   }
